refactor(analysis): decompose cache.go and reduce import fan-in
